mypage 폴더 만들어서 dto 다 옮김
