+++ src/main/java/site/metacoding/miniproject1/service/MyPageService.java
@@ -6,19 +6,19 @@ import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
 import site.metacoding.miniproject1.domain.applicationStatus.ApplicationStatusDao;
-import site.metacoding.miniproject1.web.dto.response.CompanyInfoDto;
-import site.metacoding.miniproject1.web.dto.response.InfoAllDto;
-import site.metacoding.miniproject1.web.dto.response.InfoCountDto;
-import site.metacoding.miniproject1.web.dto.response.RequestsDto;
-import site.metacoding.miniproject1.web.dto.response.RequestsInfoDto;
-import site.metacoding.miniproject1.web.dto.response.StatusAllDto;
-import site.metacoding.miniproject1.web.dto.response.StatusCountDto;
-import site.metacoding.miniproject1.web.dto.response.StatusFinalDto;
-import site.metacoding.miniproject1.web.dto.response.StatusFinalInfoDto;
-import site.metacoding.miniproject1.web.dto.response.StatusInfoDto;
-import site.metacoding.miniproject1.web.dto.response.StatusWaitingDto;
-import site.metacoding.miniproject1.web.dto.response.StatusWaitingInfoDto;
-import site.metacoding.miniproject1.web.dto.response.UserInfoDto;
+import site.metacoding.miniproject1.web.dto.response.mypage.CompanyInfoDto;
+import site.metacoding.miniproject1.web.dto.response.mypage.InfoAllDto;
+import site.metacoding.miniproject1.web.dto.response.mypage.InfoCountDto;
+import site.metacoding.miniproject1.web.dto.response.mypage.RequestsDto;
+import site.metacoding.miniproject1.web.dto.response.mypage.RequestsInfoDto;
+import site.metacoding.miniproject1.web.dto.response.mypage.StatusAllDto;
+import site.metacoding.miniproject1.web.dto.response.mypage.StatusCountDto;
+import site.metacoding.miniproject1.web.dto.response.mypage.StatusFinalDto;
+import site.metacoding.miniproject1.web.dto.response.mypage.StatusFinalInfoDto;
+import site.metacoding.miniproject1.web.dto.response.mypage.StatusInfoDto;
+import site.metacoding.miniproject1.web.dto.response.mypage.StatusWaitingDto;
+import site.metacoding.miniproject1.web.dto.response.mypage.StatusWaitingInfoDto;
+import site.metacoding.miniproject1.web.dto.response.mypage.UserInfoDto;
 
 @RequiredArgsConstructor
 @Service
@@ -30,14 +30,13 @@ public class MyPageService {
         List<UserInfoDto> userInfoDtos = applicationStatusDao.findUser(id);
         List<InfoCountDto> infoCountDtos = applicationStatusDao.findInfoCounts(id);
         List<CompanyInfoDto> companyInfoDtos = applicationStatusDao.findCompany(id);
-        if(companyInfoDtos.size() == 0) {
-        	companyInfoDtos = applicationStatusDao.findCompanyIfNull();
+        if (companyInfoDtos.size() == 0) {
+            companyInfoDtos = applicationStatusDao.findCompanyIfNull();
         }
         List<StatusCountDto> statusCountDtos = applicationStatusDao.findCounts(id);
         InfoAllDto infoAllDto = new InfoAllDto(userInfoDtos, infoCountDtos, companyInfoDtos, statusCountDtos);
         return infoAllDto;
     }
-
 
     public StatusAllDto viewAll(Integer id, String keyword) {
         List<StatusCountDto> statusCountDtos = applicationStatusDao.findCounts(id);
@@ -46,7 +45,6 @@ public class MyPageService {
         return statusAllDto;
     }
 
-
     public StatusWaitingDto viewWaiting(Integer id, String keyword) {
         List<StatusCountDto> statusCountDtos = applicationStatusDao.findCounts(id);
         List<StatusWaitingInfoDto> statusWaitingInfoDtos = applicationStatusDao.findWaiting(id, keyword);
@@ -54,14 +52,12 @@ public class MyPageService {
         return statusWaitingDto;
     }
 
-
     public StatusFinalDto viewFinal(Integer id, String keyword) {
         List<StatusCountDto> statusCountDtos = applicationStatusDao.findCounts(id);
         List<StatusFinalInfoDto> statusFinalInfoDtos = applicationStatusDao.findFinal(id, keyword);
         StatusFinalDto statusFinalDto = new StatusFinalDto(statusCountDtos, statusFinalInfoDtos);
         return statusFinalDto;
     }
-
 
     public RequestsDto viewRequests(Integer id, String keyword) {
         List<RequestsInfoDto> requestsInfoDtos = applicationStatusDao.findRequest(id, keyword);
