+++ src/main/java/site/metacoding/miniproject1/web/dto/response/mypage/InfoAllDto.java
@@ -1,4 +1,4 @@
-package site.metacoding.miniproject1.web.dto.response;
+package site.metacoding.miniproject1.web.dto.response.mypage;
 
 import java.util.List;
