+++ src/main/java/site/metacoding/miniproject1/web/dto/response/mypage/StatusCountDto.java
@@ -1,4 +1,4 @@
-package site.metacoding.miniproject1.web.dto.response;
+package site.metacoding.miniproject1.web.dto.response.mypage;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
