+++ src/main/java/site/metacoding/miniproject1/web/dto/response/mypage/StatusFinalDto.java
@@ -1,0 +1,24 @@
+package site.metacoding.miniproject1.web.dto.response.mypage;
+
+import java.util.List;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Setter
+@Getter
+public class StatusFinalDto {
+
+  private String keyword;
+  private List<StatusCountDto> statusCountDtos;
+  private List<StatusFinalInfoDto> statusFinalInfoDtos;
+
+  public StatusFinalDto(List<StatusCountDto> statusCountDtos, List<StatusFinalInfoDto> statusFinalInfoDtos) {
+    this.statusCountDtos = statusCountDtos;
+    this.statusFinalInfoDtos = statusFinalInfoDtos;
+  }
+}
