+++ src/main/java/site/metacoding/miniproject1/web/dto/response/mypage/StatusFinalInfoDto.java
@@ -1,4 +1,4 @@
-package site.metacoding.miniproject1.web.dto.response;
+package site.metacoding.miniproject1.web.dto.response.mypage;
 
 import java.sql.Timestamp;
 
@@ -11,16 +11,13 @@ import lombok.Setter;
 @AllArgsConstructor
 @Setter
 @Getter
-public class StatusInfoDto {
-
+public class StatusFinalInfoDto {
     private Integer id;
     private Integer wantedId;
     private Integer resumeId;
     private Integer state;
-    private Integer positionCodeId;
     private Timestamp created;
     private String positionCodeName;
     private Integer companyId;
     private String companyName;
-    private Integer userId;
 }
