+++ src/main/java/site/metacoding/miniproject1/web/dto/response/mypage/StatusInfoDto.java
@@ -1,4 +1,4 @@
-package site.metacoding.miniproject1.web.dto.response;
+package site.metacoding.miniproject1.web.dto.response.mypage;
 
 import java.sql.Timestamp;
 
@@ -11,7 +11,8 @@ import lombok.Setter;
 @AllArgsConstructor
 @Setter
 @Getter
-public class StatusWaitingInfoDto {
+public class StatusInfoDto {
+
     private Integer id;
     private Integer wantedId;
     private Integer resumeId;
@@ -21,4 +22,5 @@ public class StatusWaitingInfoDto {
     private String positionCodeName;
     private Integer companyId;
     private String companyName;
+    private Integer userId;
 }
