+++ src/main/java/site/metacoding/miniproject1/web/dto/response/mypage/UserInfoDto.java
@@ -1,7 +1,6 @@
-package site.metacoding.miniproject1.web.dto.response;
+package site.metacoding.miniproject1.web.dto.response.mypage;
 
 import java.sql.Blob;
-import java.sql.Timestamp;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
