+++ src/main/java/site/metacoding/miniproject1/websocket/MessageController.java
@@ -8,8 +8,8 @@ import org.springframework.messaging.simp.annotation.SendToUser;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.util.HtmlUtils;
 
-import site.metacoding.miniproject1.web.dto.response.Message;
-import site.metacoding.miniproject1.web.dto.response.PushMessage;
+import site.metacoding.miniproject1.web.dto.response.mypage.Message;
+import site.metacoding.miniproject1.web.dto.response.mypage.PushMessage;
 
 @Controller
 public class MessageController {
